modified scripts and deleted bin folders for online installation
Former-commit-id: d95aad18d6a0f1cd4faedad2456c02be710efc2f

+++ dagent/script/build_dagent.sh
@@ -13,7 +13,10 @@ export BUILD_TIME_DAGENT=$(date +%s)
 
 go build -mod vendor -tags ssloff -ldflags "-X dagent/src/routers/m9k/api/dagent.GitCommit=$GIT_COMMIT_DAGENT -X dagent/src/routers/m9k/api/dagent.BuildTime=$BUILD_TIME_DAGENT"
 
-rm -rf ./bin
+if [ -d "bin" ]
+then
+   rm -rf ./bin
+fi
 mkdir bin
 mv dagent ./bin
 cp config.yaml ./bin
